Separate MarketoApi class by target (Lead or ActivityLog)
